fix mscpool for large msc NRs
Use the proper type that can handle the entire range of MSC numbers we
allow on the VTY, we have:
 #define MSC_NR_RANGE "<0-1000>"

Before this patch, MSC pool round robin would glitch around for any
'msc' numbers 'msc 256' thru 'msc 1000'.

+++ src/osmo-bsc/gsm_08_08.c
@@ -182,7 +182,7 @@ static struct bsc_msc_data *bsc_find_msc(struct gsm_subscriber_connection *conn,
 	struct bsc_msc_data *msc_target = NULL;
 	struct bsc_msc_data *msc_round_robin_next = NULL;
 	struct bsc_msc_data *msc_round_robin_first = NULL;
-	uint8_t round_robin_next_nr;
+	unsigned int round_robin_next_nr;
 	int16_t nri_v = -1;
 	bool is_null_nri = false;
